GGRC-7524/Update integration tests for Risks to use ext_api_client
Tests for Risks and Controls objects have used obsolete api client as external. Modify tests to use external_api_helper.
1) Replace api_helper with ext_api_helper
2) Modify syntax and body of requests
3) Move external user tests to external_app package
4) Add add_person_global_role to ggrc.utils.helpers
5) Change test_assessment_url_and_status_put to use add_person_global_role
